requirements.txt
bcrypt==3.2.0
capstone==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colored-traceback==0.3.0
cryptography==36.0.0
idna==3.3
intervaltree==3.1.0
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
paramiko==2.8.1
plumbum==1.7.1
psutil==5.8.0
pwntools==4.7.0
pycparser==2.21
pyelftools==0.27
Pygments==2.10.0
PyNaCl==1.4.0
pyparsing==3.0.6
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.26.0
ROPGadget==6.6
rpyc==5.0.1
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.0rc5.post1
urllib3==1.26.7
